Extract first-login profile completion gate into a controller concern.
Move profile completion checks out of ApplicationController into a dedicated concern to improve separation of concerns while preserving onboarding behavior and coverage.
